pyproject.toml
[project]
name = "al-pr-review"
version = "1.0.0"
description = "Extract review rules from AngelList PR review threads"
license = "MIT"
requires-python = ">=3.10"
dependencies = ["pyyaml"]
